Add Plug.Conn url to context
Exception reports do not currently include the url of the request. This change
makes Airbrakex.Plug include the url in the context as does the Rails Airbrake
gem.

I'd like to use Plug.Conn.request_url, but I think the dependency on Plug would
need to be optional. Also, request_url is only available on Plug 1.5, which may
force folks to update a lot of other things to get 1.5 working. This commit
avoids the dependency problem for now by copying in the request_url code.